+++ src/Application/Common/Interfaces/ISpoonAdapter.cs
@@ -5,5 +5,5 @@ namespace RecipeApi.Application.Common.Interfaces;
 public interface ISpoonAdapter
 {
     Task<List<Recipe>> GetRandomRecipesAsync(string query);
-    Task<List<ExtendedIngredient>> GetRecipeIngredientsAsync(int recipeId);
+    Task<List<ExtendedIngredient>> GetRecipeIngredientsAsync(long recipeId);
 }
